Moved iterable wrappers to Wrappers namespace
