+++ src/Wrappers/IterableWrapper.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CardinalCollections;
+namespace CardinalCollections\Wrappers;
 
 use Iterator;
